data_manager/requirements.txt
flask==3.0.2
flask-cors==4.0.0
gevent==24.2.1
gunicorn==21.2.0
python-dotenv==1.0.1
minio==7.2.7